Update to use new Page components added along with Table
