+++ packages/edit-site/src/style.scss
@@ -8,9 +8,9 @@
 @import "./components/global-styles/screen-revisions/style.scss";
 @import "./components/header-edit-mode/style.scss";
 @import "./components/header-edit-mode/document-actions/style.scss";
-@import "./components/library/style.scss";
 @import "./components/list/style.scss";
 @import "./components/page/style.scss";
+@import "./components/page-library/style.scss";
 @import "./components/table/style.scss";
 @import "./components/sidebar-edit-mode/style.scss";
 @import "./components/sidebar-edit-mode/page-panels/style.scss";
